pyproject.toml
# contents of pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epitope_aligner"
version = "0.0.3dev"

requires-python = ">=3.10"

dependencies = [
  "pandas",
  "biopython",
  "scipy",
  "matplotlib"
]